Migrate more warnings to generated highlights
